third_party/WebKit/LayoutTests/cssom/cssfloat-cssom.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="target"></div>  

<script>
addEventListener("load", function() {
  test(function() {
    assert_true('cssFloat' in getComputedStyle(target), 'cssFloat is in CSSStyleDeclaration');
    assert_equals(target.style.cssFloat, '');
    target.style.cssFloat = 'right';
    assert_equals(target.style.cssFloat, 'right');

    var CSSXGetComputedStyleQueries = 566;
    assert_false(internals.isUseCounted(document, CSSXGetComputedStyleQueries));
  }, 'Testing CSSFloat');
}, false);
</script>
